Tested Debit Card. OK. Added Credit card service

+++ backend/cards-service/src/main/java/com/digitalbank/cards_service/models/CreditCard.java
@@ -15,8 +15,8 @@ import java.util.Date;
 public class CreditCard extends Card {
     private String cardName;
     private String cardType;
-    private Double creditLimit;
-    private Double availableLimit;
+    private BigInteger creditLimit;
+    private BigInteger availableLimit;
     private String cardClass;
     @Temporal(TemporalType.DATE)
     private Date billGenerationDate;
